requirements.txt
email
discord
random
yfinance
time
pandas
numpy
matplotlib
datetime
plotly.graph_objs
kaleido